Split src/preview/launch.ts into launch/ subdir (warren-62a7 / pl-9088 step 9)
src/preview/launch.ts (712 lines) and launch.test.ts (995 lines) split
into a same-name subdirectory so each module stays well under the
500-line file-size ratchet:

  - launch/index.ts       — public API re-exports + module walkthrough
  - launch/types.ts       — PreviewSidecarsClient, LaunchPreviewInput,
                            LaunchPreviewResult, LaunchFailureReason, all
                            DEFAULT_* tunables
  - launch/probe.ts       — tcpConnectOnce (phase 1, raw TCP) + probeOnce
                            (phase 2, HTTP) health-check primitives
  - launch/helpers.ts     — captureFailureTail, safeDeleteSidecar,
                            composeFailureMessage, truncate, defaultSleep,
                            defaultSidecarEnv
  - launch/setup.ts       — runSetupStep (warren-d9e7 setup pre-step)
  - launch/orchestrate.ts — launchPreview state machine
  - launch/url.ts         — formatPreviewUrl + loadPreviewLaunchConfigFromEnv

Tests split alongside (each well under the 500-line cap):

  - launch/orchestrate.test.ts (377)
  - launch/probe-loop.test.ts  (255)  — warren-9b15 two-phase probe describe
  - launch/setup.test.ts       (158)  — warren-d9e7 setup pre-step describe
  - launch/probe.test.ts       (73)   — tcpConnectOnce direct Bun.connect tests
  - launch/url.test.ts         (58)
  - launch/test-helpers.ts     (182)  — shared fakes + setupLaunchEnv

Call sites updated (server/main.ts, server/bridges.ts,
runs/reap/{preview,preview.test,types}.ts, preview/eviction/types.ts).

biome.json: removed launch.test.ts from noExcessiveLinesPerFunction
override; replaced launch.ts with launch/orchestrate.ts + launch/setup.ts
in the noExcessiveCognitiveComplexity override (both still ~16).

scripts/file-size-budgets.json: removed both old entries — no new
grandfathering needed.

bun run check:all is green end-to-end.

Refs: warren-62a7, pl-9088 step 9, warren-1b72.
