+++ scripts/file-size-budgets.json
@@ -20,8 +20,6 @@
 		"src/plan-runs/coordinator.ts": 587,
 		"src/plots/aggregate.test.ts": 847,
 		"src/plots/aggregate.ts": 506,
-		"src/preview/launch.test.ts": 995,
-		"src/preview/launch.ts": 712,
 		"src/projects/manage.test.ts": 763,
 		"src/projects/refresh.test.ts": 709,
 		"src/registry/refresh.test.ts": 1041,
